refactored CombinatorialComplex and CombinatorialConformation and Tests
